refactor: consolidate command palette open onto event bus, remove React context
Replace useOpenAddProjectCommandPalette() context hook with direct
openCommandPalette({ open: 'add-project' }) bus calls across all consumers.
Move isCommandPaletteOpen() into commandPaletteBus.ts. Delete
commandPaletteContext.tsx and dead newThreadPickerBus.ts.
